+++ src/collective/contact/plonegroup/utils.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+from operator import attrgetter
 from operator import methodcaller
 from zope.schema.vocabulary import SimpleVocabulary, SimpleTerm
 from plone import api
@@ -54,4 +55,8 @@ def voc_selected_org_suffix_users(org_uid, suffixes, first_member=None):
                 value=member.getUserName(),  # login
                 token=member.getId(),  # id
                 title=member.getUser().getProperty('fullname') or member.getUserName()))  # title
+    if first_member is None:
+        terms.sort(key=attrgetter('title'))
+    else:
+        terms[1:] = sorted(terms[1:], key=attrgetter('title'))
     return SimpleVocabulary(terms)